==> 0_setup_env.ipynb <==
## 0 setup environment
install all necessary library and create folder for later use

**if your test system is Linux, please change pip to pip3**

!pip install msgpack
!pip install vaderSentiment
!pip install textblob
!pip install nltk
!pip install matplotlib==2.0.2
!pip install pandas
!pip install seaborn

import nltk
nltk.download('stopwords')
nltk.download('punkt')

# optional
# import nltk
# nltk.download('all')

# create folders
import os, sys

if os.path.isdir('extract'):  
    pass  
else:  
    os.mkdir('extract')

if os.path.isdir('zip_folder'):  
    pass  
else:  
    os.mkdir('zip_folder')

if os.path.isdir('obj'):  
    pass  
else:  
    os.mkdir('obj')
print('Done')
