==> 1_unzip.ipynb <==
## 1. unzip
Unzip files to /extract folder

**input**  :zip files at /zip_folder  
**output** :json files at /extract folder

import json
import sqlite3
import os, sys
import zipfile

'''Specify the relatively path'''
path = "./extract"
zip_folder = "./zip_folder/"

'''function un_zip: unzip the given zip file, input file path'''
def un_zip(file_name, unzip_file):  
    zip_file = zipfile.ZipFile(file_name)  
    for names in zip_file.namelist():
        zip_file.extract(names,unzip_file)
    zip_file.close()

'''function execute_unzip: unzip the zip file at once to /extract folder'''
def execute_unzip():
    print('found',len(zip_file_list),'zips')
    for files in zip_file_list:
        print('unzip ',files,' complete')
        un_zip(zip_folder+str(files), path)
    print('Done')

# create a list that only contains zip file
zip_file_list = []

#select only zip files, append the names into list
for file in os.listdir(zip_folder):
    if os.path.splitext(file)[1][1:] == "zip":
        zip_file_list.append(file)

execute_unzip()
