==> 4_conversation_table.ipynb <==
## 4. create conversation table
Insert conversation table to database core.db  

**input**: conversation pickle file  
**output** : database conversation table

import pandas as pd
import sqlite3
import time
import pickle

file = 'core.db'
conn = sqlite3.connect(file)

'''
This way you can load the pickle file into python.

It contains our beloveth conversations with super badass scores.
'''

pickle_conversation = pickle.load(open("obj/conversations_with_scores.p", "rb"))

airlines = ["56377143", "106062176", "18332190", "22536055", "124476322", "26223583", "2182373406", 
                "38676903", "1542862735", "253340062", "218730857", "45621423", "20626359"]

count = 0
conv = []
tweet_id =[]
for conversation_index in range(len(pickle_conversation)):
        for tweet_index in range(len(pickle_conversation[conversation_index])):
            tweet_id.append(pickle_conversation[conversation_index][tweet_index][0])
            conv.append(count)
        count += 1

frame = pd.DataFrame(data = tweet_id, columns= ["tweet_id"])
frame["conversation_id"] = conv

'''adds the conversation table to the database'''

frame.to_sql("conversation", conn, if_exists= "replace")

print('done')
