==> 7_hex_plot_with_R_code.ipynb <==
## 7. hex plot
Code for creating the last poster plot, to visualize the sentiment change after airline replys to negaitve tweets  

**input**: word_sentiment_change csv file  
**output**: hex plot

## 7-2. R code
Generate data for R code to run t test of user sentiment **before** and **after** airline replys

**input**: word_sentiment_change csv file  
**output**: mean and standard deviation of **before** and **after** replys

import pandas as pd


%matplotlib inline
import matplotlib.pyplot as plt
import seaborn as sns
sns.set()  # set Seaborn defaults
plt.rcParams['figure.figsize'] = 15, 10  # default hor./vert. size of plots, in inches
plt.rcParams['lines.markeredgewidth'] = 1  # to fix issue with seaborn box plots; needed after import seaborn
import numpy as np

'''opens the csv file that contains the sentiment before and after AA replies to tweets that contain 
"cancel", "delay" or "connection" '''
file_dir = 'obj/'
file = file_dir + "word_sentiment_change.csv" 
filtered_comb = pd.read_csv(file)
filtered_comb.head()

sns.set(font_scale = 1.5)
sns.axes_style("white")
sns.jointplot(x=filtered_comb["first"], y=filtered_comb["last"] , kind="hex", color = "#7b0323", size = 10, xlim=(-0.9,-0.01));

#plt.savefig('sentiment_change.png');
#the red diagonal line on the poster has been added afterwards, to make it easier to see

## R Code variables 
After executing the following cells, the output can be simply copied into R, since these variables will then be used for the tsum.test function.

print("n <-", len(filtered_comb))

print("first_mean <- ",filtered_comb["first"].mean())
print("first_sd <- ",filtered_comb["first"].std())

print("last_mean <-",filtered_comb["last"].mean()) 
print("last_sd <-",filtered_comb["last"].std())

print('Done')
